+++ project/src/test/java/org/apache/modele/TestTournee.java
@@ -466,5 +466,298 @@ public class TestTournee {
 				"\n" + 
 				"", tournee.exporterRoute());
 	}
+	
+	@Test(timeout = 1000)
+	public void testAjoutSimple() throws ParserConfigurationException, SAXException, IOException, ExceptionXML {
+		// Creation des objets plan et demande
+		File xml = new File("src/test/java/org/apache/modele/fichiers/DLpetit5.xml");
+		File planxml = new File("src/test/java/org/apache/modele/fichiers/planLyonPetit.xml");
+		PlanDeVille plan = new PlanDeVille();
+		Deserialisateur.chargerPlanDeVilleFichier(plan, planxml);
+		DemandeDeLivraison demande = new DemandeDeLivraison();
+		Deserialisateur.chargerDemandeLivraisonFichier(demande, plan, xml);
+
+		// Calcul tournee
+		Tournee tournee = new Tournee();
+		tournee.setEntrepot(demande.getEntrepot());
+		tournee.calculerTournee(plan, demande, 10000);
+
+		// Copie des chemins initiaux
+		List<Chemin> anciensChemins = new ArrayList<Chemin>();
+		for (Chemin c : tournee.getChemins()) {
+			anciensChemins.add(c);
+		}
+
+		// On recupère la livraison suivante et la livraison précédente
+
+		Livraison livraisonPre = tournee.getLivraison(1);
+		Livraison livraisonSuiv = tournee.getLivraison(2);
+
+		Intersection intersectionLiv = plan.getIntersectionById((long) 251171098);
+		Livraison livraisonNouv = new Livraison(intersectionLiv, 600);
+
+		// Calcule des nouveaux chemins
+		Chemin chemin1 = tournee.calculerNouveauChemin(plan, livraisonPre.getLieuDeLivraison(),
+				livraisonNouv.getLieuDeLivraison());
+		Chemin chemin2 = tournee.calculerNouveauChemin(plan, livraisonNouv.getLieuDeLivraison(),
+				livraisonSuiv.getLieuDeLivraison());
+
+		// Mise à jour de tournee
+		tournee.ajouterListeLivraison(livraisonNouv, 2);
+		tournee.supprimerChemin(1);
+		tournee.ajouterChemin(chemin1, 1);
+		tournee.ajouterChemin(chemin2, 2);
+
+		// Verification de l'ordre et des intersection a livrer
+		assertEquals(25321357, (long) tournee.getLivraisonsOrdonnees().get(0).getLieuDeLivraison().getIdNoeud());
+		assertEquals(1860559399, (long) tournee.getLivraisonsOrdonnees().get(1).getLieuDeLivraison().getIdNoeud());
+		assertEquals(251171098, (long) tournee.getLivraisonsOrdonnees().get(2).getLieuDeLivraison().getIdNoeud());
+		assertEquals(25303807, (long) tournee.getLivraisonsOrdonnees().get(3).getLieuDeLivraison().getIdNoeud());
+		assertEquals(26155540, (long) tournee.getLivraisonsOrdonnees().get(4).getLieuDeLivraison().getIdNoeud());
+		assertEquals(29003879, (long) tournee.getLivraisonsOrdonnees().get(5).getLieuDeLivraison().getIdNoeud());
+
+		// Verification de l'ordre des tronçons
+		List<Chemin> chemins = new ArrayList<Chemin>();
+		chemins = tournee.getChemins();
+		assertEquals(anciensChemins.get(0), chemins.get(0));
+		assertEquals(chemin1, chemins.get(1));
+		assertEquals(chemin2, chemins.get(2));
+		assertEquals(anciensChemins.get(2), chemins.get(3));
+		assertEquals(anciensChemins.get(3), chemins.get(4));
+		assertEquals(anciensChemins.get(4), chemins.get(5));
+		
+		int ancienneDuree = tournee.getDureeTourneeSecondes();
+		 
+		tournee.calculerDureeTotale();
+		 
+		int nouvelleDuree = tournee.getDureeTourneeSecondes();
+		 
+		assertTrue(ancienneDuree < nouvelleDuree); //La duree a bien ete changee
+		// WARNING: Test de non regression -> cette valeur n'est pas 100% sûr
+		assertEquals(4980, nouvelleDuree);
+	}
+
+	@Test(timeout = 1000)
+	public void testAjoutApresEntrepot() throws ParserConfigurationException, SAXException, IOException, ExceptionXML {
+		// Creation des objets plan et demande
+		File xml = new File("src/test/java/org/apache/modele/fichiers/DLpetit5.xml");
+		File planxml = new File("src/test/java/org/apache/modele/fichiers/planLyonPetit.xml");
+		PlanDeVille plan = new PlanDeVille();
+		Deserialisateur.chargerPlanDeVilleFichier(plan, planxml);
+		DemandeDeLivraison demande = new DemandeDeLivraison();
+		Deserialisateur.chargerDemandeLivraisonFichier(demande, plan, xml);
+
+		// Calcul tournee
+		Tournee tournee = new Tournee();
+		tournee.setEntrepot(demande.getEntrepot());
+		tournee.calculerTournee(plan, demande, 10000);
+
+		// Copie des chemins initiaux
+		List<Chemin> anciensChemins = new ArrayList<Chemin>();
+		for (Chemin c : tournee.getChemins()) {
+			anciensChemins.add(c);
+		}
+
+		// On recupère la livraison suivante et la livraison précédente
+		Livraison livraisonSuiv = tournee.getLivraison(0);
+
+		Intersection intersectionLiv = plan.getIntersectionById((long) 251171098);
+		Livraison livraisonNouv = new Livraison(intersectionLiv, 600);
+
+		// Calcule des nouveaux chemins
+		Chemin chemin1 = tournee.calculerNouveauChemin(plan, tournee.getEntrepot().getLieuDeLivraison(),
+				livraisonNouv.getLieuDeLivraison());
+		Chemin chemin2 = tournee.calculerNouveauChemin(plan, livraisonNouv.getLieuDeLivraison(),
+				livraisonSuiv.getLieuDeLivraison());
+
+		// Mise à jour de tournee
+		tournee.ajouterListeLivraison(livraisonNouv, 1);
+		tournee.supprimerChemin(0);
+		tournee.ajouterChemin(chemin1, 0);
+		tournee.ajouterChemin(chemin2, 1);
+
+		// Verification de l'ordre et des intersection a livrer
+		assertEquals(25321357, (long) tournee.getLivraisonsOrdonnees().get(0).getLieuDeLivraison().getIdNoeud());
+		assertEquals(251171098, (long) tournee.getLivraisonsOrdonnees().get(1).getLieuDeLivraison().getIdNoeud());
+		assertEquals(1860559399, (long) tournee.getLivraisonsOrdonnees().get(2).getLieuDeLivraison().getIdNoeud());
+		assertEquals(25303807, (long) tournee.getLivraisonsOrdonnees().get(3).getLieuDeLivraison().getIdNoeud());
+		assertEquals(26155540, (long) tournee.getLivraisonsOrdonnees().get(4).getLieuDeLivraison().getIdNoeud());
+		assertEquals(29003879, (long) tournee.getLivraisonsOrdonnees().get(5).getLieuDeLivraison().getIdNoeud());
+
+		// Verification de l'ordre des tronçons
+		List<Chemin> chemins = new ArrayList<Chemin>();
+		chemins = tournee.getChemins();
+		assertEquals(chemin1, chemins.get(0));
+		assertEquals(chemin2, chemins.get(1));
+		assertEquals(anciensChemins.get(1), chemins.get(2));
+		assertEquals(anciensChemins.get(2), chemins.get(3));
+		assertEquals(anciensChemins.get(3), chemins.get(4));
+		assertEquals(anciensChemins.get(4), chemins.get(5));
+	}
+
+	@Test(timeout = 1000)
+	public void testAjoutAvantEntrepot() throws ParserConfigurationException, SAXException, IOException, ExceptionXML {
+		// Creation des objets plan et demande
+		File xml = new File("src/test/java/org/apache/modele/fichiers/DLpetit5.xml");
+		File planxml = new File("src/test/java/org/apache/modele/fichiers/planLyonPetit.xml");
+		PlanDeVille plan = new PlanDeVille();
+		Deserialisateur.chargerPlanDeVilleFichier(plan, planxml);
+		DemandeDeLivraison demande = new DemandeDeLivraison();
+		Deserialisateur.chargerDemandeLivraisonFichier(demande, plan, xml);
+
+		// Calcul tournee
+		Tournee tournee = new Tournee();
+		tournee.setEntrepot(demande.getEntrepot());
+		tournee.calculerTournee(plan, demande, 10000);
+
+		// Copie des chemins initiaux
+		List<Chemin> anciensChemins = new ArrayList<Chemin>();
+		for (Chemin c : tournee.getChemins()) {
+			anciensChemins.add(c);
+		}
+
+		// On recupère la livraison suivante et la livraison précédente
+		Livraison livraisonPre = tournee.getLivraison(3);
+
+		Intersection intersectionLiv = plan.getIntersectionById((long) 251171098);
+		Livraison livraisonNouv = new Livraison(intersectionLiv, 600);
+
+		// Calcule des nouveaux chemins
+		Chemin chemin1 = tournee.calculerNouveauChemin(plan, livraisonPre.getLieuDeLivraison(),
+				livraisonNouv.getLieuDeLivraison());
+		Chemin chemin2 = tournee.calculerNouveauChemin(plan, livraisonNouv.getLieuDeLivraison(),
+				tournee.getEntrepot().getLieuDeLivraison());
+
+		// Mise à jour de tournee
+		tournee.ajouterListeLivraison(livraisonNouv, 5);
+		tournee.supprimerChemin(4);
+		tournee.ajouterChemin(chemin1, 4);
+		tournee.ajouterChemin(chemin2, 5);
+
+		// Verification de l'ordre et des intersection a livrer
+		assertEquals(25321357, (long) tournee.getLivraisonsOrdonnees().get(0).getLieuDeLivraison().getIdNoeud());
+		assertEquals(1860559399, (long) tournee.getLivraisonsOrdonnees().get(1).getLieuDeLivraison().getIdNoeud());
+		assertEquals(25303807, (long) tournee.getLivraisonsOrdonnees().get(2).getLieuDeLivraison().getIdNoeud());
+		assertEquals(26155540, (long) tournee.getLivraisonsOrdonnees().get(3).getLieuDeLivraison().getIdNoeud());
+		assertEquals(29003879, (long) tournee.getLivraisonsOrdonnees().get(4).getLieuDeLivraison().getIdNoeud());
+		assertEquals(251171098, (long) tournee.getLivraisonsOrdonnees().get(5).getLieuDeLivraison().getIdNoeud());
+
+		// Verification de l'ordre des tronçons
+		List<Chemin> chemins = new ArrayList<Chemin>();
+		chemins = tournee.getChemins();
+		assertEquals(anciensChemins.get(0), chemins.get(0));
+		assertEquals(anciensChemins.get(1), chemins.get(1));
+		assertEquals(anciensChemins.get(2), chemins.get(2));
+		assertEquals(anciensChemins.get(3), chemins.get(3));
+		assertEquals(chemin1, chemins.get(4));
+		assertEquals(chemin2, chemins.get(5));
+	}
+
+	@Test(timeout = 1000)
+	public void testAjoutMutiples() throws ParserConfigurationException, SAXException, IOException, ExceptionXML {
+		// Creation des objets plan et demande
+		File xml = new File("src/test/java/org/apache/modele/fichiers/DLpetit5.xml");
+		File planxml = new File("src/test/java/org/apache/modele/fichiers/planLyonPetit.xml");
+		PlanDeVille plan = new PlanDeVille();
+		Deserialisateur.chargerPlanDeVilleFichier(plan, planxml);
+		DemandeDeLivraison demande = new DemandeDeLivraison();
+		Deserialisateur.chargerDemandeLivraisonFichier(demande, plan, xml);
+
+		// Calcul tournee
+		Tournee tournee = new Tournee();
+		tournee.setEntrepot(demande.getEntrepot());
+		tournee.calculerTournee(plan, demande, 10000);
+
+		// Copie des chemins initiaux
+		List<Chemin> anciensChemins = new ArrayList<Chemin>();
+		for (Chemin c : tournee.getChemins()) {
+			anciensChemins.add(c);
+		}
+
+		// On recupere la livraison suivante et la livraison precedente
+		Livraison livraisonPre = tournee.getLivraison(0);
+		Livraison livraisonSuiv = tournee.getLivraison(1);
+
+		Intersection intersectionLiv = plan.getIntersectionById((long) 251171098);
+		Livraison livraisonNouv = new Livraison(intersectionLiv, 600);
+
+		// Calcule des nouveaux chemins
+		Chemin chemin1 = tournee.calculerNouveauChemin(plan, livraisonPre.getLieuDeLivraison(),
+				livraisonNouv.getLieuDeLivraison());
+		Chemin chemin2 = tournee.calculerNouveauChemin(plan, livraisonNouv.getLieuDeLivraison(),
+				livraisonSuiv.getLieuDeLivraison());
+				
+		// Mise a jour de tournee
+		tournee.ajouterListeLivraison(livraisonNouv, 2);
+		tournee.supprimerChemin(1);
+		tournee.ajouterChemin(chemin1, 1);
+		tournee.ajouterChemin(chemin2, 2);
+
+		// Verification de l'ordre et des intersections a livrer
+		assertEquals(25321357, (long) tournee.getLivraisonsOrdonnees().get(0).getLieuDeLivraison().getIdNoeud());
+		assertEquals(1860559399, (long) tournee.getLivraisonsOrdonnees().get(1).getLieuDeLivraison().getIdNoeud());
+		assertEquals(251171098, (long) tournee.getLivraisonsOrdonnees().get(2).getLieuDeLivraison().getIdNoeud());
+		assertEquals(25303807, (long) tournee.getLivraisonsOrdonnees().get(3).getLieuDeLivraison().getIdNoeud());
+		assertEquals(26155540, (long) tournee.getLivraisonsOrdonnees().get(4).getLieuDeLivraison().getIdNoeud());
+		assertEquals(29003879, (long) tournee.getLivraisonsOrdonnees().get(5).getLieuDeLivraison().getIdNoeud());
+
+		// Verification de l'ordre des troncons
+		List<Chemin> chemins = new ArrayList<Chemin>();
+		chemins = tournee.getChemins();
+		assertEquals(anciensChemins.get(0), chemins.get(0));
+		assertEquals(chemin1, chemins.get(1));
+		assertEquals(chemin2, chemins.get(2));
+		assertEquals(anciensChemins.get(2), chemins.get(3));
+		assertEquals(anciensChemins.get(3), chemins.get(4));
+		assertEquals(anciensChemins.get(4), chemins.get(5));
+
+		/*************************** Nouvel ajout ************************************/
+		tournee.calculerDureeTotale();
+
+		anciensChemins = new ArrayList<Chemin>();
+		for (Chemin c : tournee.getChemins()) {
+			anciensChemins.add(c);
+		}
+
+		// On recupere la livraison suivante et la livraison precedente
+		livraisonPre = tournee.getLivraison(1);
+		livraisonSuiv = tournee.getLivraison(2);
+
+		intersectionLiv = plan.getIntersectionById((long) 1383027249);
+		livraisonNouv = new Livraison(intersectionLiv, 600);
+
+		// Calcul des nouveaux chemins
+		chemin1 = tournee.calculerNouveauChemin(plan, livraisonPre.getLieuDeLivraison(),
+				livraisonNouv.getLieuDeLivraison());
+		chemin2 = tournee.calculerNouveauChemin(plan, livraisonNouv.getLieuDeLivraison(),
+				livraisonSuiv.getLieuDeLivraison());
+
+		// Mise a jour de tournee
+		tournee.ajouterListeLivraison(livraisonNouv, 2);
+		tournee.supprimerChemin(1);
+		tournee.ajouterChemin(chemin1, 1);
+		tournee.ajouterChemin(chemin2, 2);
+
+		// Verification de l'ordre et des intersections a livrer
+		assertEquals(25321357, (long) tournee.getLivraisonsOrdonnees().get(0).getLieuDeLivraison().getIdNoeud());
+		assertEquals(1860559399, (long) tournee.getLivraisonsOrdonnees().get(1).getLieuDeLivraison().getIdNoeud());
+		assertEquals(1383027249, (long) tournee.getLivraisonsOrdonnees().get(2).getLieuDeLivraison().getIdNoeud());
+		assertEquals(251171098, (long) tournee.getLivraisonsOrdonnees().get(3).getLieuDeLivraison().getIdNoeud());
+		assertEquals(25303807, (long) tournee.getLivraisonsOrdonnees().get(4).getLieuDeLivraison().getIdNoeud());
+		assertEquals(26155540, (long) tournee.getLivraisonsOrdonnees().get(5).getLieuDeLivraison().getIdNoeud());
+		assertEquals(29003879, (long) tournee.getLivraisonsOrdonnees().get(6).getLieuDeLivraison().getIdNoeud());
+
+		// Verification de l'ordre des tronçons
+		chemins = new ArrayList<Chemin>();
+		chemins = tournee.getChemins();
+		assertEquals(anciensChemins.get(0), chemins.get(0));
+		assertEquals(chemin1, chemins.get(1));
+		assertEquals(chemin2, chemins.get(2));
+		assertEquals(anciensChemins.get(2), chemins.get(3));
+		assertEquals(anciensChemins.get(3), chemins.get(4));
+		assertEquals(anciensChemins.get(4), chemins.get(5));
+		assertEquals(anciensChemins.get(5), chemins.get(6));
+
+	}
 
 }
